requirements.txt
streamlit~=1.40.1
pandas~=2.0.3
mysql-connector-python~=9.0.0
pymongo~=4.10.1